requirements.txt
Flask
pytesseract
Pillow
langdetect
translatepy
gunicorn
